+++ src/factorial.h
@@ -5,8 +5,8 @@
 #ifndef __RAMANUJANFACTORIAL__
 #define __RAMANUJANFACTORIAL__
 
-inline long double _ramanujan(unsigned int n) {
-  long double f= sqrt(M_PI) * pow(n, n) * exp(-n) * pow((8 * pow(n, 3) + 4 * pow(n, 2) + n + 1 / 30), 1 / 6);
+inline long double _ramanujan(double n) {
+  long double f = sqrt(M_PI) * pow(n, n) * exp(-n) * pow((8.0 * pow(n, 3.0) + 4.0 * pow(n, 2.0) + n + 1.0 / 30.0), 1.0 / 6.0);
   return f;
 }
 
@@ -16,9 +16,8 @@ inline long double _ramanujan(unsigned int n) {
 #ifndef __STIRLINGFACTORIAL__
 #define __STIRLINGFACTORIAL__
 
-inline long double _stirling(unsigned int n) {
+inline long double _stirling(double n) {
   long double f = sqrt((2.0 * n + 1.0 / 3.0) * M_PI) * pow(n, n) * exp(-n);
-
   return f;
 }
 
@@ -28,7 +27,7 @@ inline long double _stirling(unsigned int n) {
 #ifndef __STIRLINGLNFACTORIAL__
 #define __STIRLINGLNFACTORIAL__
 
-inline long double _stirlingln(unsigned int n) {
+inline long double _stirlingln(double n) {
   long double f = n * log(n) - n + (1.0 / 6.0) * log(n * (1.0 + 4.0 * n * (1.0 + 2.0 * n))) + 0.5 * log(M_PI);
 
   return f;
@@ -41,9 +40,9 @@ inline long double _stirlingln(unsigned int n) {
 #define __FACTORIAL__
 
 inline long int _factorial(unsigned int n) {
-  short int fac = 1;
+  unsigned long int fac = 1;
 
-  for(unsigned int i = n; i >= 1; i--) {
+  for (unsigned int i = n; i >= 1; i--) {
     fac = fac * i;
   }
